requirements.txt
opencv-contrib-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
qrcode[pil]==7.4.2